+++ packages/dom-capture/src/browser/captureFrame.js
@@ -3,6 +3,7 @@ const defaultDomProps = require('./defaultDomProps');
 const getBackgroundImageUrl = require('./getBackgroundImageUrl');
 const getImageSizes = require('./getImageSizes');
 const genXpath = require('./genXpath');
+const isInlineFrame = require('./isInlineFrame');
 const absolutizeUrl = require('./absolutizeUrl');
 const makeGetBundledCssFromCssText = require('./getBundledCssFromCssText');
 const parseCss = require('./parseCss');
@@ -18,6 +19,7 @@ async function captureFrame(
   {styleProps, rectProps, ignoredTagNames} = defaultDomProps,
   doc = document,
   addStats = false,
+  fetchTimeLimit = 30000,
 ) {
   const performance = {total: {}, prefetchCss: {}, doCaptureFrame: {}, waitForImages: {}};
   function startTime(obj) {
@@ -36,7 +38,7 @@ async function captureFrame(
   const separator = '-----';
 
   startTime(performance.prefetchCss);
-  const prefetchAllCss = makePrefetchAllCss(makeFetchCss(fetch));
+  const prefetchAllCss = makePrefetchAllCss(makeFetchCss(fetch, {fetchTimeLimit}));
   const getCssFromCache = await prefetchAllCss(doc);
   endTime(performance.prefetchCss);
 
@@ -108,7 +110,7 @@ async function captureFrame(
     };
   }
 
-  function doCaptureFrame(frameDoc) {
+  function doCaptureFrame(frameDoc, baseUrl = frameDoc.location.href) {
     const bgImages = new Set();
     let bundledCss = '';
     const ret = captureNode(frameDoc.documentElement);
@@ -119,7 +121,7 @@ async function captureFrame(
     function captureNode(node) {
       const {bundledCss: nodeCss, unfetchedResources: nodeUnfetched} = captureNodeCss(
         node,
-        frameDoc.location.href,
+        baseUrl,
       );
       bundledCss += nodeCss;
       if (nodeUnfetched) for (const elem of nodeUnfetched) unfetchedResources.add(elem);
@@ -196,7 +198,9 @@ async function captureFrame(
       }
       try {
         if (doc) {
-          obj.childNodes = [doCaptureFrame(el.contentDocument)];
+          obj.childNodes = [
+            doCaptureFrame(doc, isInlineFrame(el) ? el.baseURI : doc.location.href),
+          ];
         } else {
           markFrameAsCors();
         }