+++ packages/eyes-sdk-core/lib/runner/EyesRunner.js
@@ -1,16 +1,18 @@
 'use strict'
 
 const {GeneralUtils} = require('@applitools/eyes-common')
-const getScmInfo = require('../getScmInfo')
 
 class EyesRunner {
   constructor() {
     /** @type {Eyes[]} */
     this._eyesInstances = []
+    this._getBatchInfo = undefined
   }
 
-  makeGetBatchInfo(fetchBatchInfo) {
+  attachEyes(eyes, serverConnector) {
+    this._eyesInstances.push(eyes)
     if (!this._getBatchInfo) {
+      const fetchBatchInfo = serverConnector.batchInfo.bind(serverConnector)
       this._getBatchInfo = GeneralUtils.cachify(fetchBatchInfo)
     }
   }
@@ -19,14 +21,8 @@ class EyesRunner {
     if (this._getBatchInfo) {
       return this._getBatchInfo(batchId)
     } else {
-      throw new Error(
-        'Eyes runner could not get batch info since makeGetBatchInfo was not called before',
-      )
+      throw new Error('Eyes runner could not get batch info since attachEyes was not called before')
     }
-  }
-
-  async getScmInfoWithCache(...args) {
-    return getScmInfo(...args)
   }
 
   /**