+++ packages/eyes-webdriverio-4/src/EyesWDIO.js
@@ -70,11 +70,8 @@ class EyesWDIO extends EyesBase {
    **/
   constructor(serverUrl, isDisabled = false, runner = new ClassicRunner()) {
     super(serverUrl, isDisabled, new Configuration())
-    /** @type {EyesRunner} */ this._runner = runner
-
-    this._runner._eyesInstances.push(this)
-    this._runner.makeGetRenderingInfo(this._serverConnector.renderInfo.bind(this._serverConnector))
-    this._runner.makeGetBatchInfo(this._serverConnector.batchInfo.bind(this._serverConnector))
+    this._runner = runner
+    this._runner.attachEyes(this, this._serverConnector)
 
     /** @type {EyesWebDriver} */
     this._driver = undefined