+++ packages/eyes-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## Unreleased
+
+## [3.19.0]
+
+- batchInfo has generating batch id info
+
 ## [3.18.5]
 
 - fix isFeatureFlagOn and isFeatureFlagOff