+++ packages/eyes-protractor/CHANGELOG.md
@@ -1,7 +1,6 @@
 # Change Log
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+## Unreleased
 
 ## [4.8.59](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-protractor@4.8.58...@applitools/eyes-protractor@4.8.59) (2019-12-02)