+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## Unreleased
+
+## 8.0.1
+
+getUserSetBatchId() can now run after generating batch id
+
+## 8.0.0
+
+Moved server connector makers to runners
+
 ## 7.0.0
 
 - move all runners to eyes-sdk-core