+++ packages/eyes-webdriverio-5/CHANGELOG.md
@@ -1,7 +1,12 @@
 # Change Log
 
-## 5.9.12
+## Unreleased
 
+## 5.9.13
+- branching base commit support
+- make functions in runners 
+
+## 5.9.12
 - Send stitching service URL to visual grid [Trello 212](https://trello.com/c/Sqh6k2VV)
 
 ## 5.9.9 - 2020-02-09